Kafka/requirements.txt
kafka-python==2.0.2
pyspark==3.5.1
pandas==2.2.2
numpy==1.26.4
seaborn==0.13.2
matplotlib==3.8.4
scikit-learn==1.4.2